Extract Loader, add loading delay to random question
